Cleanup some code around RenderFunctions
