+++ src/React.Core/IReactComponent.cs
@@ -9,7 +9,6 @@
 
 using System;
 using System.IO;
-using React.RenderFunctions;
 
 namespace React
 {
@@ -57,7 +56,7 @@ namespace React
 		/// <param name="exceptionHandler">A custom exception handler that will be called if a component throws during a render. Args: (Exception ex, string componentName, string containerId)</param>
 		/// <param name="renderFunctions"></param>
 		/// <returns>HTML</returns>
-		string RenderHtml(bool renderContainerOnly = false, bool renderServerOnly = false, Action<Exception, string, string> exceptionHandler = null, IRenderFunctions renderFunctions = null);
+		string RenderHtml(bool renderContainerOnly = false, bool renderServerOnly = false, Action<Exception, string, string> exceptionHandler = null, RenderFunctions renderFunctions = null);
 
 		/// <summary>
 		/// Renders the JavaScript required to initialise this component client-side. This will
@@ -77,7 +76,7 @@ namespace React
 		/// <param name="exceptionHandler">A custom exception handler that will be called if a component throws during a render. Args: (Exception ex, string componentName, string containerId)</param>
 		/// <param name="renderFunctions"></param>
 		/// <returns>HTML</returns>
-		void RenderHtml(TextWriter writer, bool renderContainerOnly = false, bool renderServerOnly = false, Action<Exception, string, string> exceptionHandler = null, IRenderFunctions renderFunctions = null);
+		void RenderHtml(TextWriter writer, bool renderContainerOnly = false, bool renderServerOnly = false, Action<Exception, string, string> exceptionHandler = null, RenderFunctions renderFunctions = null);
 
 		/// <summary>
 		/// Renders the JavaScript required to initialise this component client-side. This will
