+++ src/React.Core/StyledComponentsRenderFunctions.cs
@@ -1,27 +1,27 @@
 using System;
 
-namespace React.RenderFunctions
+namespace React
 {
 	/// <summary>
-	/// 
+	/// Render functions for styled components
 	/// </summary>
-	public class StyledComponentsFunctions : RenderFunctionsBase
+	public class StyledComponentsFunctions : RenderFunctions
 	{
 		private Action<string> _onPostRender;
 
 		/// <summary>
-		/// 
+		///
 		/// </summary>
 		/// <param name="renderFunctions"></param>
 		/// <param name="onPostRender"></param>
-		public StyledComponentsFunctions(IRenderFunctions renderFunctions = null, Action<string> onPostRender = null)
+		public StyledComponentsFunctions(RenderFunctions renderFunctions = null, Action<string> onPostRender = null)
 			: base(renderFunctions)
 		{
 			_onPostRender = onPostRender;
 		}
 
 		/// <summary>
-		/// 
+		///
 		/// </summary>
 		/// <param name="executeJs"></param>
 		protected override void PreRenderCore(Func<string, string> executeJs)
@@ -30,7 +30,7 @@ namespace React.RenderFunctions
 		}
 
 		/// <summary>
-		/// 
+		///
 		/// </summary>
 		/// <param name="componentToRender"></param>
 		/// <returns></returns>
@@ -40,7 +40,7 @@ namespace React.RenderFunctions
 		}
 
 		/// <summary>
-		/// 
+		///
 		/// </summary>
 		/// <param name="executeJs"></param>
 		protected override void PostRenderCore(Func<string, string> executeJs)
