+++ src/React.Sample.Router.CoreMvc/Views/Home/Index.cshtml
@@ -1,6 +1,5 @@
 @using React.AspNet
 @using React.Router
-@using React.RenderFunctions
 
 @{
 	Layout = "_Layout";
